#239 moved the queries around
