+++ JavaFXUserInterface/src/main/java/Controllers/AddSystemTypeController.java
@@ -4,13 +4,13 @@ import Utilities.TextConstants;
 import Utilities.UIUtilities;
 import app.item.AssetType;
 import app.item.AssetTypeParameter;
+import external.AssetTypeDAOImpl;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
 import javafx.scene.control.TextFormatter;
-import local.AssetTypeDAOImpl;
 
 import java.net.URL;
 import java.util.ArrayList;
