+++ JavaFXUserInterface/src/main/java/Controllers/SystemInfoController.java
@@ -4,6 +4,10 @@ import Utilities.UIUtilities;
 import app.item.Asset;
 import app.item.AssetAttribute;
 import app.item.Measurement;
+import external.AssetDAOImpl;
+import external.AssetTypeDAOImpl;
+import external.AttributeDAOImpl;
+import external.ModelDAOImpl;
 import javafx.animation.Animation;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
@@ -22,10 +26,6 @@ import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.Pane;
 import javafx.scene.text.Text;
 import javafx.util.Duration;
-import local.AssetDAOImpl;
-import local.AssetTypeDAOImpl;
-import local.AttributeDAOImpl;
-import local.ModelDAOImpl;
 import rul.assessment.AssessmentController;
 
 import java.net.URL;
