+++ JavaFXUserInterface/src/main/java/external/AttributeDAO.java
@@ -1,4 +1,4 @@
-package local;
+package external;
 
 import app.item.Measurement;
