+++ JavaFXUserInterface/src/main/java/external/AttributeDAOImpl.java
@@ -1,4 +1,4 @@
-package local;
+package external;
 
 import app.item.Measurement;
