+++ JavaFXUserInterface/src/main/java/external/ModelDAO.java
@@ -1,0 +1,22 @@
+/*
+  Interface for the ModelDAO object
+
+  @author Paul Micu
+  @version 1.0
+  @last_edit 12/27/2020
+ */
+package external;
+
+import app.item.TrainedModel;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public interface ModelDAO {
+
+    TrainedModel getModelsByAssetTypeID(String assetTypeID);
+
+    String getModelNameFromModelID(int modelID);
+
+    TrainedModel createTrainedModelFromResultSet(ResultSet rs, boolean withModel) throws SQLException;
+}
