+++ JavaFXUserInterface/src/test/java/AddSystemTypeControllerTest.java
@@ -16,7 +16,8 @@ import org.junit.Test;
 import org.testfx.api.FxToolkit;
 import org.testfx.framework.junit.ApplicationTest;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 public class AddSystemTypeControllerTest extends ApplicationTest {
     private Scene scene;
@@ -27,7 +28,7 @@ public class AddSystemTypeControllerTest extends ApplicationTest {
         FXMLLoader fxmlLoader = new FXMLLoader(SystemTypeInfoController.class.getResource("/AddSystemType.fxml"));
         Parent root = fxmlLoader.load();
         scene = new Scene(root);
-        addSystemTypeController = (AddSystemTypeController) fxmlLoader.getController();
+        addSystemTypeController = fxmlLoader.getController();
         stage.setTitle("CBMS");
         stage.setScene(scene);
         stage.show();
