+++ JavaFXUserInterface/src/test/java/SystemInfoControllerTest.java
@@ -1,6 +1,7 @@
 import Controllers.SystemInfoController;
 import Controllers.SystemsController;
 import app.item.Asset;
+import external.AssetDAOImpl;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
@@ -13,7 +14,6 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.text.Text;
 import javafx.stage.Stage;
-import local.AssetDAOImpl;
 import org.junit.After;
 import org.junit.Test;
 import org.testfx.api.FxAssert;
@@ -23,7 +23,8 @@ import org.testfx.matcher.base.NodeMatchers;
 
 import java.util.ArrayList;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 public class SystemInfoControllerTest extends ApplicationTest {
 
@@ -35,7 +36,7 @@ public class SystemInfoControllerTest extends ApplicationTest {
         FXMLLoader fxmlLoader = new FXMLLoader(SystemsController.class.getResource("/SystemInfo.fxml"));
         Parent root = fxmlLoader.load();
         scene = new Scene(root);
-        systemInfoController = (SystemInfoController) fxmlLoader.getController();
+        systemInfoController = fxmlLoader.getController();
 //        Asset asset = new Asset();
 //        asset.setId(123);
 //        asset.setLocation("location");
